@@ -68,7 +68,8 @@
     "marked": "^0.3.3",
     "moment": "^2.8.2",
     "query-string": "^4.1.0",
-    "react": "^0.13.3",
+    "react": "^0.14.8",
+    "react-dom": "^0.14.8",
     "scriptjs": "^2.5.7",
     "shortid": "^2.2.6",
     "source-map-support": "^0.2.7",
